Size

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int Int_type;
 long int Long_int_type;
 long long int Long_long_int_type;
 float Float_type;
-printf("Size of a char: %ld byte(s)\n",sizeof(Char_type));
-printf("Size of a int: %ld byte(s)\n",sizeof(Int_type));
-printf("Size of a long int: %ld byte(s)\n",sizeof(Long_int_type));
-printf("Size of a long long int: %ld byte(s)\n",sizeof(Long_long_int_type));
-printf("Size of a float: %ld byte(s)\n",sizeof(Float_type));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(Char_type));
+printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(Int_type));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(Long_int_type));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(Long_long_int_type));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(Float_type));
 return (0);
 }
